updating sync
