+++ lua/grug-far/engine/ripgrep.lua
@@ -2,8 +2,10 @@ local fetchCommandOutput = require('grug-far/engine/fetchCommandOutput')
 local parseResults = require('grug-far/engine/ripgrep/parseResults')
 local fetchFilesWithMatches = require('grug-far/engine/ripgrep/fetchFilesWithMatches')
 local replaceInMatchedFiles = require('grug-far/engine/ripgrep/replaceInMatchedFiles')
+local syncChangedFiles = require('grug-far/engine/ripgrep/syncChangedFiles')
 local getArgs = require('grug-far/engine/ripgrep/getArgs')
 local colors = require('grug-far/engine/ripgrep/colors')
+local utils = require('grug-far/utils')
 
 --- are we replacing matches with the empty string?
 ---@param args string[]
@@ -22,6 +24,20 @@ local function isEmptyStringReplace(args)
   end
 
   return true
+end
+
+--- are we doing a multiline search and replace?
+---@param args string[]
+---@return boolean
+local function isMultilineSearchReplace(args)
+  local multilineFlags = { '--multiline', '-U', '--multiline-dotall' }
+  for _, arg in ipairs(args) do
+    if utils.isBlacklistedFlag(arg, multilineFlags) then
+      return true
+    end
+  end
+
+  return false
 end
 
 ---@type GrugFarEngine
@@ -112,6 +128,30 @@ local RipgrepEngine = {
     })
 
     return abort
+  end,
+
+  sync = function(params)
+    local on_finish = params.on_finish
+
+    local args = getArgs(params.inputs, params.options, {})
+    if not args then
+      on_finish(nil, nil, 'sync cannot work with the current arguments!')
+      return
+    end
+
+    if isMultilineSearchReplace(args) then
+      on_finish(nil, nil, 'sync disabled for multline search/replace!')
+      return
+    end
+
+    return syncChangedFiles({
+      options = params.options,
+      report_progress = function(count)
+        params.report_progress({ type = 'update_count', count = count })
+      end,
+      on_finish = params.on_finish,
+      changedFiles = params.changedFiles,
+    })
   end,
 }
